# Saving boot time during ProcessesWithContext call (linux).