On Linux, the system boot time is used when calculating process creation times (expressed as number of seconds since the epoch). Obtaining the boot time involves parsing the /proc/stat file, which can incur siginificant overhead on systems with many CPU cores.

Permanently caching the boot time is not an option since it can change when the system clock is updated (e.g. ntpd).

In this changeset we will save the boot time during one invocation of the ProcessesWithContext function (listing all running processes).