@bjandras 2023-08-07
This addresses issues #1070 and #1283.

To my knowledge, obtaining the boot time from /proc/stat is the only reliable
method. Alternative of subtracting the uptime from the current time is not
reliable since it requires obtaining these two times simultaneously.