@bjandras 2023-08-07
Simple benchmark (just calling `process.Processes()` for `testing.B.N` times:

before:
```
goos: linux
goarch: amd64
pkg: testtest
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkBTime-4             196           6110134 ns/op
PASS
ok      testtest        2.272s
```

with change:
```
goos: linux
goarch: amd64
pkg: testtest
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkBTime-4             445           2677522 ns/op
PASS
ok      testtest        1.480s
```

(depends heavily on the number of processes running)