@shirou 2023-08-16
Ah, I am sorry that I have misled you.  What I meant was that caching in context should be done by the Application that uses gopsutil, not in the gopsutil, because it is a Library. To use a different term, we are unable to accept this PR.