@bjandras 2023-08-16
@shirou I don't understand -- caching is done by the application, the library only uses the cached value when available.