@shirou 2023-08-17
We do not want to add more context values. And also, bootTime can jump if time gets adjusted, described in [this comment](https://github.com/shirou/gopsutil/issues/1070#issuecomment-842512782).